requirements-runtime.txt
# Runtime dependencies for Lambda functions
boto3>=1.34.0
requests>=2.31.0
tenacity>=8.2.0
jinja2>=3.1.0
pyyaml>=6.0.0
httpx>=0.27.0
